requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.32.4
selenium~=4.34.2